Update namespace and add StringNullOrEmptyToVisibilityConverter
Updated the namespace for the converters XML namespace declaration in App.xaml from MyApp.Converters to MyApp.Utils.

Added a new file StringNullOrEmptyToVisibilityConverter.cs which defines the StringNullOrEmptyToVisibilityConverter class in the MyApp.Utils namespace. This class implements the IValueConverter interface to convert a string to a boolean value based on whether the string is null or empty. The ConvertBack method is not implemented and throws a NotImplementedException.