Add ColumnSpacing to Grid and remove Margin from Buttons
The Grid layout in BooksListingPage.xaml now includes a ColumnSpacing property set to 5, adding a 5-unit space between each column. Additionally, the Margin attribute for the Delete and Download buttons in Grid.Column="1" and Grid.Column="2" was removed, eliminating the 8-unit margin on the left and right sides of these buttons.